@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.12.3
+// - protoc             v3.18.1
 // source: api/micro/moment/v1/comment.proto
 
 package v1
@@ -27,11 +27,13 @@ type CommentServiceClient interface {
 	// 删除评论
 	DeleteComment(ctx context.Context, in *DeleteCommentRequest, opts ...grpc.CallOption) (*DeleteCommentReply, error)
 	// 查看评论
+	//
 	//	rpc GetComment (GetCommentRequest) returns (GetCommentReply) {
 	//		option (google.api.http) = {
 	//			get: "/v1/comments/{id}"
 	//		};
 	//	};
+	//
 	// 热门评论列表
 	ListHotComment(ctx context.Context, in *ListCommentRequest, opts ...grpc.CallOption) (*ListCommentReply, error)
 	// 最新评论列表
@@ -113,11 +115,13 @@ type CommentServiceServer interface {
 	// 删除评论
 	DeleteComment(context.Context, *DeleteCommentRequest) (*DeleteCommentReply, error)
 	// 查看评论
+	//
 	//	rpc GetComment (GetCommentRequest) returns (GetCommentReply) {
 	//		option (google.api.http) = {
 	//			get: "/v1/comments/{id}"
 	//		};
 	//	};
+	//
 	// 热门评论列表
 	ListHotComment(context.Context, *ListCommentRequest) (*ListCommentReply, error)
 	// 最新评论列表
