@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.12.3
+// - protoc             v3.18.1
 // source: api/micro/relation/v1/relation.proto
 
 package v1
