@@ -387,8 +387,8 @@ func (m *LoginRequest) GetPassword() string {
 }
 
 type LoginReply struct {
-	Uid   string `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
-	Token string `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
+	Uid         string `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
+	AccessToken string `protobuf:"bytes,2,opt,name=access_token,json=accessToken,proto3" json:"access_token,omitempty"`
 }
 
 func (m *LoginReply) Reset()         { *m = LoginReply{} }
@@ -431,16 +431,16 @@ func (m *LoginReply) GetUid() string {
 	return ""
 }
 
-func (m *LoginReply) GetToken() string {
+func (m *LoginReply) GetAccessToken() string {
 	if m != nil {
-		return m.Token
+		return m.AccessToken
 	}
 	return ""
 }
 
 type LogoutRequest struct {
-	Id    int64  `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
-	Token string `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
+	Id          int64  `protobuf:"varint,1,opt,name=id,proto3" json:"id,omitempty"`
+	AccessToken string `protobuf:"bytes,2,opt,name=access_token,json=accessToken,proto3" json:"access_token,omitempty"`
 }
 
 func (m *LogoutRequest) Reset()         { *m = LogoutRequest{} }
@@ -483,9 +483,9 @@ func (m *LogoutRequest) GetId() int64 {
 	return 0
 }
 
-func (m *LogoutRequest) GetToken() string {
+func (m *LogoutRequest) GetAccessToken() string {
 	if m != nil {
-		return m.Token
+		return m.AccessToken
 	}
 	return ""
 }
@@ -816,7 +816,7 @@ var xxx_messageInfo_UpdateUserReply proto.InternalMessageInfo
 
 type UpdatePasswordRequest struct {
 	Id              string `protobuf:"bytes,1,opt,name=id,proto3" json:"id,omitempty"`
-	OldPassword     string `protobuf:"bytes,2,opt,name=old_password,json=oldPassword,proto3" json:"old_password,omitempty"`
+	Password        string `protobuf:"bytes,2,opt,name=password,proto3" json:"password,omitempty"`
 	NewPassword     string `protobuf:"bytes,3,opt,name=new_password,json=newPassword,proto3" json:"new_password,omitempty"`
 	ConfirmPassword string `protobuf:"bytes,4,opt,name=confirm_password,json=confirmPassword,proto3" json:"confirm_password,omitempty"`
 }
@@ -861,9 +861,9 @@ func (m *UpdatePasswordRequest) GetId() string {
 	return ""
 }
 
-func (m *UpdatePasswordRequest) GetOldPassword() string {
+func (m *UpdatePasswordRequest) GetPassword() string {
 	if m != nil {
-		return m.OldPassword
+		return m.Password
 	}
 	return ""
 }
@@ -1119,89 +1119,90 @@ func init() {
 func init() { proto.RegisterFile("api/micro/user/v1/user.proto", fileDescriptor_23a8e11f455382c5) }
 
 var fileDescriptor_23a8e11f455382c5 = []byte{
-	// 1305 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xb4, 0x57, 0x4d, 0x6c, 0x1b, 0xc5,
-	0x17, 0x8f, 0xbd, 0x89, 0x9d, 0x3c, 0x3b, 0x8e, 0x33, 0x6d, 0x52, 0x77, 0x9b, 0xda, 0xee, 0xfe,
-	0xff, 0x6d, 0x8d, 0x51, 0xbc, 0x8d, 0x4b, 0x45, 0xe5, 0x72, 0x71, 0xd2, 0x50, 0x95, 0xba, 0x29,
-	0xda, 0xa4, 0xaa, 0x84, 0x90, 0xa2, 0xb5, 0x77, 0x62, 0x8f, 0x6a, 0xef, 0x6c, 0x77, 0xc7, 0xae,
-	0x22, 0xc4, 0x05, 0x24, 0x0e, 0x20, 0x21, 0x24, 0x84, 0x04, 0x57, 0xce, 0x48, 0x3d, 0x21, 0x50,
-	0x2f, 0x5c, 0x38, 0x70, 0xac, 0xc4, 0x85, 0x53, 0x84, 0xda, 0x42, 0x51, 0x8e, 0x48, 0x9c, 0xb8,
-	0xa0, 0x99, 0xfd, 0xec, 0xd6, 0x09, 0x2d, 0x52, 0x2f, 0x99, 0xb7, 0x6f, 0x7e, 0xef, 0xfd, 0xde,
-	0xcc, 0xfb, 0xf0, 0x04, 0x96, 0x74, 0x8b, 0xa8, 0x03, 0xd2, 0xb1, 0xa9, 0x3a, 0x74, 0xb0, 0xad,
-	0x8e, 0x56, 0xc4, 0x5a, 0xb3, 0x6c, 0xca, 0x28, 0x9a, 0xd7, 0x2d, 0x52, 0x13, 0xbb, 0x35, 0xa1,
-	0x1d, 0xad, 0xc8, 0x4b, 0x5d, 0x4a, 0xbb, 0x7d, 0xac, 0x72, 0x3b, 0xdd, 0x34, 0x29, 0xd3, 0x19,
-	0xa1, 0xa6, 0xe3, 0x1a, 0xc8, 0x85, 0x2e, 0xed, 0x52, 0x55, 0xc8, 0xed, 0xe1, 0x8e, 0xca, 0xbf,
-	0xbc, 0x9d, 0x13, 0xd4, 0xc2, 0xa6, 0x6e, 0x91, 0xd1, 0xf9, 0x67, 0xcd, 0x94, 0xbf, 0x25, 0x98,
-	0xbc, 0xe9, 0x60, 0x1b, 0x2d, 0x42, 0x92, 0x18, 0x85, 0x44, 0x39, 0x51, 0x99, 0x59, 0x4d, 0xed,
-	0xef, 0x95, 0x92, 0xc4, 0xd0, 0x92, 0xc4, 0x40, 0x15, 0x98, 0xe6, 0x01, 0x98, 0xfa, 0x00, 0x17,
-	0x92, 0x62, 0x37, 0xbb, 0xbf, 0x57, 0x0a, 0x74, 0x5a, 0x20, 0xa1, 0x12, 0x4c, 0xe1, 0x81, 0x4e,
-	0xfa, 0x05, 0x49, 0xc0, 0x66, 0xf6, 0xf7, 0x4a, 0xae, 0x42, 0x73, 0x17, 0x0e, 0xb0, 0x7a, 0xd4,
-	0xc4, 0x85, 0xc9, 0x10, 0x20, 0x14, 0x9a, 0xbb, 0xa0, 0xb3, 0x30, 0xdd, 0xa7, 0x5d, 0x62, 0x6e,
-	0xeb, 0xac, 0x30, 0x55, 0x4e, 0x54, 0x24, 0x97, 0xcb, 0xd7, 0x69, 0x69, 0x21, 0x35, 0x19, 0x6a,
-	0x42, 0xca, 0x61, 0x3a, 0x1b, 0x3a, 0x85, 0x54, 0x39, 0x51, 0xc9, 0xd5, 0x4f, 0xd6, 0x9e, 0xb9,
-	0xae, 0xda, 0xa6, 0x00, 0x6c, 0xed, 0x5a, 0x78, 0x15, 0xf6, 0xf7, 0x4a, 0x9e, 0x81, 0xe6, 0xad,
-	0xfc, 0x5c, 0x26, 0xe9, 0xdc, 0x16, 0xe7, 0x4a, 0x87, 0xe7, 0xf2, 0x75, 0x5a, 0x20, 0x21, 0x05,
-	0x52, 0xfa, 0x48, 0x67, 0xba, 0x5d, 0x98, 0x16, 0x38, 0xe1, 0xcd, 0xd5, 0x68, 0xde, 0xca, 0x03,
-	0xea, 0x62, 0xd3, 0xc0, 0x76, 0x61, 0xe6, 0xc0, 0x80, 0xae, 0x08, 0x40, 0x18, 0x90, 0x6b, 0xa0,
-	0x79, 0x2b, 0x0f, 0xa8, 0x4d, 0x6c, 0xd6, 0x33, 0xf4, 0xdd, 0x02, 0x84, 0x01, 0xf9, 0x3a, 0x2d,
-	0x90, 0xd0, 0x71, 0x90, 0xda, 0x84, 0x16, 0x32, 0x02, 0x94, 0xde, 0xdf, 0x2b, 0xf1, 0x4f, 0x8d,
-	0xff, 0x41, 0xcb, 0x00, 0x1d, 0x1b, 0xeb, 0x0c, 0x1b, 0xfc, 0x0e, 0xb3, 0xe2, 0x0e, 0x73, 0xfb,
-	0x7b, 0xa5, 0x88, 0x56, 0x9b, 0xf1, 0xe4, 0x26, 0x53, 0xb6, 0x61, 0x4e, 0xc3, 0x5d, 0xe2, 0x30,
-	0x6c, 0x6b, 0xf8, 0xce, 0x10, 0x3b, 0x0c, 0xc9, 0x91, 0x7c, 0x8b, 0x6a, 0x88, 0x64, 0xf8, 0xa8,
-	0x9f, 0x61, 0x51, 0x08, 0x7e, 0x5a, 0x65, 0x98, 0xb6, 0x74, 0xc7, 0xb9, 0x4b, 0x6d, 0xc3, 0x4d,
-	0xbd, 0x16, 0x7c, 0x2b, 0x97, 0x60, 0x36, 0x24, 0xb0, 0xfa, 0xbb, 0x28, 0x17, 0x94, 0x99, 0x24,
-	0xca, 0x4b, 0x8e, 0x97, 0x57, 0x48, 0xa7, 0xbc, 0x0b, 0xd9, 0x16, 0x4f, 0xf8, 0xcb, 0x09, 0xed,
-	0x35, 0x00, 0xcf, 0x3b, 0x8f, 0x2b, 0x0f, 0xd2, 0xd0, 0xaf, 0x7f, 0x8d, 0x8b, 0xdc, 0x23, 0xa3,
-	0xb7, 0xb1, 0xe9, 0x7b, 0x14, 0x1f, 0xca, 0x05, 0x98, 0x6d, 0xd1, 0x2e, 0x1d, 0x32, 0x3f, 0xa8,
-	0xf8, 0x81, 0xc6, 0x9b, 0xcd, 0x42, 0xc6, 0x37, 0xb3, 0xfa, 0xbb, 0x8a, 0x0e, 0xf3, 0x6b, 0x22,
-	0x09, 0xbc, 0xf5, 0x5e, 0xce, 0xf1, 0x36, 0x61, 0x2e, 0x4a, 0xf1, 0x82, 0x77, 0x1f, 0x12, 0x4a,
-	0x11, 0x42, 0xe5, 0xaf, 0x24, 0xcc, 0xdf, 0xb4, 0x8c, 0x58, 0xe0, 0xa1, 0xdf, 0x99, 0xff, 0xe6,
-	0x97, 0x6b, 0x23, 0x93, 0xc1, 0x1f, 0x07, 0xc7, 0xe3, 0xe3, 0x20, 0x1c, 0x00, 0x17, 0x5e, 0x68,
-	0x00, 0x04, 0x4d, 0x2f, 0xc7, 0x9b, 0x3e, 0xd2, 0xe6, 0x8b, 0x4f, 0xb7, 0x79, 0xd0, 0xda, 0x17,
-	0x5e, 0xa8, 0xb5, 0x83, 0x76, 0x96, 0xe3, 0xed, 0x1c, 0x69, 0xe0, 0x7c, 0xa4, 0x81, 0xdd, 0xbe,
-	0x3d, 0x09, 0x30, 0x14, 0xf7, 0x2a, 0xfa, 0x76, 0x56, 0x1c, 0x76, 0xc6, 0xd3, 0x34, 0x99, 0x32,
-	0x0f, 0x73, 0xd1, 0x6b, 0xe7, 0x25, 0xf4, 0x65, 0x02, 0x16, 0x5c, 0xdd, 0xdb, 0x5e, 0xca, 0x0f,
-	0x4a, 0xc7, 0x29, 0xc8, 0xd2, 0xbe, 0xb1, 0x1d, 0x54, 0x8a, 0x9b, 0x92, 0x0c, 0xed, 0x1b, 0xbe,
-	0x25, 0x87, 0x98, 0xf8, 0xee, 0x76, 0xac, 0x98, 0x32, 0x26, 0xbe, 0x1b, 0x40, 0x5e, 0x81, 0x7c,
-	0x87, 0x9a, 0x3b, 0xc4, 0x1e, 0x84, 0x30, 0x37, 0x5b, 0x73, 0x9e, 0xde, 0x87, 0x2a, 0x0b, 0x70,
-	0x24, 0x1e, 0x19, 0x8f, 0xb8, 0x06, 0xb9, 0x2b, 0x98, 0x45, 0x0b, 0x67, 0x29, 0xf2, 0x9b, 0x93,
-	0xfd, 0x93, 0xff, 0xaa, 0xd8, 0xa4, 0xa1, 0x10, 0x43, 0xe1, 0x71, 0x2b, 0x97, 0x20, 0x1b, 0xe0,
-	0x79, 0xf9, 0xbe, 0x0a, 0x93, 0xfc, 0xbe, 0x05, 0x3e, 0x53, 0x3f, 0x36, 0x26, 0x0d, 0x02, 0x2b,
-	0x40, 0x4a, 0x05, 0x8e, 0xae, 0xea, 0xac, 0xd3, 0xf3, 0x3c, 0x38, 0x3e, 0x65, 0x1e, 0x24, 0x62,
-	0x38, 0x85, 0x44, 0x59, 0xaa, 0x48, 0x1a, 0x17, 0x95, 0x35, 0x40, 0x31, 0x24, 0x27, 0x5b, 0x86,
-	0x29, 0xee, 0xc7, 0x45, 0x1e, 0xc2, 0xe6, 0xa2, 0xaa, 0xcb, 0x00, 0x61, 0xb9, 0x21, 0x80, 0xd4,
-	0xc6, 0x0d, 0xed, 0x7a, 0xb3, 0x95, 0x9f, 0xe0, 0xf2, 0xe5, 0xf5, 0xd6, 0xfa, 0xd6, 0x7a, 0x3e,
-	0x81, 0xd2, 0x20, 0xad, 0xea, 0x66, 0x3e, 0x59, 0x55, 0x01, 0xc2, 0x92, 0x41, 0x19, 0x48, 0xdf,
-	0xdc, 0xb8, 0xb6, 0x71, 0xe3, 0xd6, 0x46, 0x7e, 0x02, 0x4d, 0xc3, 0xe4, 0xf5, 0x66, 0x8b, 0xa3,
-	0x01, 0x52, 0x6f, 0xae, 0x0b, 0x39, 0x59, 0xff, 0x66, 0x0a, 0x32, 0x9c, 0x6f, 0x13, 0xdb, 0x23,
-	0xd2, 0xc1, 0xc8, 0x84, 0x69, 0x7f, 0xae, 0x22, 0x65, 0x4c, 0x6c, 0xb1, 0xa9, 0x2e, 0x97, 0x0f,
-	0xc5, 0xf0, 0xec, 0x2c, 0x7d, 0xf0, 0xf3, 0xe3, 0xcf, 0x93, 0x8b, 0xca, 0x3c, 0x7f, 0x88, 0xe8,
-	0x43, 0xd6, 0x53, 0x6d, 0x6f, 0xbf, 0x91, 0xa8, 0xa2, 0x36, 0x4c, 0x89, 0x61, 0x89, 0x4a, 0x63,
-	0x1c, 0x45, 0x87, 0xb4, 0x7c, 0xf2, 0x60, 0x00, 0xa7, 0x39, 0x2e, 0x68, 0x8e, 0x28, 0xb9, 0x80,
-	0x46, 0xb4, 0x34, 0xe7, 0xd8, 0x81, 0x94, 0x3b, 0x23, 0x51, 0x79, 0xbc, 0x8f, 0x70, 0xea, 0xca,
-	0xc5, 0x43, 0x10, 0x9c, 0x46, 0x16, 0x34, 0x47, 0x95, 0xb9, 0x28, 0x0d, 0x1d, 0x32, 0xce, 0x83,
-	0x21, 0xed, 0xe5, 0x1a, 0x9d, 0x1a, 0xdb, 0xcb, 0xd1, 0x1a, 0x95, 0x4b, 0x87, 0x41, 0x38, 0xd5,
-	0xa2, 0xa0, 0xca, 0xa3, 0x9c, 0xff, 0x82, 0x73, 0xd4, 0xf7, 0x88, 0xf1, 0x3e, 0xba, 0x03, 0x10,
-	0xf6, 0x2c, 0xfa, 0xff, 0xb8, 0x02, 0x8a, 0x4f, 0x52, 0x59, 0xf9, 0x17, 0x54, 0xe4, 0x06, 0xe5,
-	0x18, 0x1f, 0x3f, 0xd9, 0x47, 0x09, 0xc8, 0x3d, 0xdd, 0x79, 0xa8, 0x72, 0xa0, 0xc7, 0xd8, 0xd8,
-	0x90, 0xcf, 0x3c, 0x07, 0x92, 0xf3, 0x2b, 0x82, 0x7f, 0xa9, 0x7e, 0x2c, 0xe4, 0xf7, 0x07, 0x82,
-	0x1f, 0xc8, 0xea, 0x57, 0xd2, 0xfd, 0x2b, 0x3f, 0x26, 0xd1, 0x77, 0x09, 0x48, 0x13, 0xd3, 0x59,
-	0xd6, 0x2d, 0x82, 0x4e, 0x5c, 0xc6, 0x4e, 0xc7, 0x26, 0x16, 0x7f, 0x7c, 0x96, 0x77, 0x6c, 0x3a,
-	0x28, 0x87, 0x8f, 0x51, 0xa5, 0x01, 0xd9, 0x35, 0x6a, 0x32, 0xbd, 0xc3, 0xca, 0x1b, 0x7c, 0x16,
-	0x57, 0x7b, 0x8c, 0x59, 0x4e, 0x43, 0x55, 0xbb, 0x84, 0xf5, 0x86, 0xed, 0x5a, 0x87, 0x0e, 0xd4,
-	0x2e, 0x5d, 0x16, 0xc1, 0x39, 0x6e, 0x0b, 0xa8, 0x9e, 0xe3, 0xea, 0x35, 0xc8, 0x5c, 0xbf, 0xba,
-	0x55, 0x6e, 0x91, 0x0e, 0x36, 0x1d, 0x8c, 0xde, 0x78, 0x7e, 0x53, 0xb5, 0xdd, 0xa7, 0x6d, 0x75,
-	0xa0, 0x13, 0x53, 0x6d, 0x5d, 0x5d, 0x5b, 0xdf, 0xd8, 0x5c, 0xaf, 0xa7, 0x46, 0x2b, 0xb5, 0x73,
-	0xb5, 0x73, 0xf2, 0x69, 0x58, 0xe0, 0x7e, 0x1a, 0xaa, 0xda, 0xa7, 0x1d, 0xbd, 0xdf, 0xa3, 0x0e,
-	0x6b, 0x5c, 0x3c, 0x77, 0x71, 0x05, 0x65, 0x7f, 0xff, 0xf6, 0x93, 0x3f, 0xee, 0x7d, 0xf1, 0xe4,
-	0xd3, 0x7b, 0xbf, 0x7d, 0xfd, 0xa1, 0x7c, 0x06, 0x16, 0x3d, 0x98, 0xe7, 0xb2, 0xd6, 0xa1, 0xb4,
-	0x5f, 0x23, 0x14, 0x65, 0x9f, 0xdc, 0xff, 0xe1, 0xf1, 0xf7, 0x1f, 0xbb, 0xb8, 0xea, 0xd9, 0xc6,
-	0x69, 0xf8, 0x1f, 0x40, 0x1b, 0xeb, 0x36, 0xb6, 0x9b, 0x43, 0xd6, 0x43, 0x0b, 0x70, 0x04, 0x26,
-	0xb9, 0x6d, 0x35, 0xe5, 0x6a, 0xeb, 0xd2, 0x5b, 0xb7, 0xb6, 0xea, 0x79, 0xc8, 0x3d, 0x05, 0x9b,
-	0xf8, 0xe9, 0x61, 0x31, 0xf1, 0xe0, 0x61, 0x31, 0xf1, 0xeb, 0xc3, 0x62, 0xe2, 0xb3, 0x47, 0xc5,
-	0x89, 0x07, 0x8f, 0x8a, 0x13, 0xbf, 0x3c, 0x2a, 0x4e, 0xbc, 0xf3, 0xfa, 0x73, 0x9c, 0xef, 0x99,
-	0xff, 0x30, 0x2e, 0x8d, 0x56, 0xda, 0x29, 0xf1, 0xf0, 0x3f, 0xff, 0x4f, 0x00, 0x00, 0x00, 0xff,
-	0xff, 0xe6, 0x6a, 0x09, 0x54, 0x80, 0x0c, 0x00, 0x00,
+	// 1324 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0xb4, 0x57, 0x4f, 0x68, 0x1b, 0x47,
+	0x17, 0xb7, 0xb4, 0xb6, 0x64, 0x3f, 0xc9, 0xb2, 0x3c, 0x89, 0x1d, 0x65, 0xe3, 0x48, 0xca, 0x7e,
+	0x5f, 0x12, 0x7d, 0x0e, 0xd6, 0xc6, 0x0a, 0x81, 0xe0, 0xf0, 0x1d, 0x64, 0xc7, 0x0d, 0x69, 0x15,
+	0xa7, 0xac, 0x1d, 0x02, 0xa5, 0x60, 0x46, 0xab, 0xb1, 0x34, 0x44, 0xda, 0xd9, 0xec, 0x8e, 0x14,
+	0x4c, 0xe9, 0xa5, 0x85, 0x1e, 0x5a, 0x28, 0x2d, 0xa5, 0xb7, 0x9c, 0x7a, 0x0e, 0xa4, 0x97, 0x42,
+	0x9b, 0x4b, 0x0f, 0xbd, 0xf4, 0x18, 0xe8, 0xa5, 0x27, 0x53, 0xf2, 0xa7, 0x29, 0x86, 0x5e, 0x0a,
+	0x3d, 0xf5, 0x52, 0x66, 0x76, 0x57, 0xbb, 0xde, 0x38, 0x6e, 0x5c, 0xc8, 0xc5, 0x33, 0xfb, 0xf6,
+	0x37, 0xef, 0xf7, 0xe6, 0xfd, 0xf9, 0x79, 0x05, 0x73, 0xd8, 0xa6, 0x7a, 0x8f, 0x9a, 0x0e, 0xd3,
+	0xfb, 0x2e, 0x71, 0xf4, 0xc1, 0xa2, 0x5c, 0xab, 0xb6, 0xc3, 0x38, 0x43, 0xd3, 0xd8, 0xa6, 0x55,
+	0xf9, 0xb6, 0x2a, 0xad, 0x83, 0x45, 0x75, 0xae, 0xcd, 0x58, 0xbb, 0x4b, 0x74, 0x71, 0x0e, 0x5b,
+	0x16, 0xe3, 0x98, 0x53, 0x66, 0xb9, 0xde, 0x01, 0xb5, 0xd0, 0x66, 0x6d, 0xa6, 0xcb, 0x7d, 0xb3,
+	0xbf, 0xa5, 0x8b, 0x27, 0xff, 0xcd, 0x09, 0x66, 0x13, 0x0b, 0xdb, 0x74, 0x70, 0xe1, 0xc5, 0x63,
+	0xda, 0x5f, 0x0a, 0x8c, 0xde, 0x74, 0x89, 0x83, 0x66, 0x21, 0x49, 0x5b, 0x85, 0x44, 0x39, 0x51,
+	0x99, 0x58, 0x4e, 0xed, 0xee, 0x94, 0x92, 0xb4, 0x65, 0x24, 0x69, 0x0b, 0x55, 0x60, 0x5c, 0x04,
+	0x60, 0xe1, 0x1e, 0x29, 0x24, 0xe5, 0xdb, 0xec, 0xee, 0x4e, 0x69, 0x68, 0x33, 0x86, 0x3b, 0x54,
+	0x82, 0x31, 0xd2, 0xc3, 0xb4, 0x5b, 0x50, 0x24, 0x6c, 0x62, 0x77, 0xa7, 0xe4, 0x19, 0x0c, 0x6f,
+	0x11, 0x00, 0xbb, 0xc3, 0x2c, 0x52, 0x18, 0x0d, 0x01, 0xd2, 0x60, 0x78, 0x0b, 0x3a, 0x0b, 0xe3,
+	0x5d, 0xd6, 0xa6, 0xd6, 0x26, 0xe6, 0x85, 0xb1, 0x72, 0xa2, 0xa2, 0x78, 0x5c, 0x81, 0xcd, 0x48,
+	0xcb, 0x5d, 0x9d, 0xa3, 0x3a, 0xa4, 0x5c, 0x8e, 0x79, 0xdf, 0x2d, 0xa4, 0xca, 0x89, 0x4a, 0xae,
+	0x76, 0xb2, 0xfa, 0x42, 0xba, 0xaa, 0xeb, 0x12, 0xb0, 0xb1, 0x6d, 0x93, 0x65, 0xd8, 0xdd, 0x29,
+	0xf9, 0x07, 0x0c, 0x7f, 0x15, 0xf7, 0xb2, 0xa8, 0x79, 0x5b, 0xde, 0x2b, 0x1d, 0xde, 0x2b, 0xb0,
+	0x19, 0xc3, 0x1d, 0xd2, 0x20, 0x85, 0x07, 0x98, 0x63, 0xa7, 0x30, 0x2e, 0x71, 0xd2, 0x9b, 0x67,
+	0x31, 0xfc, 0x55, 0x04, 0xd4, 0x26, 0x56, 0x8b, 0x38, 0x85, 0x89, 0x97, 0x06, 0x74, 0x55, 0x02,
+	0xc2, 0x80, 0xbc, 0x03, 0x86, 0xbf, 0x8a, 0x80, 0x9a, 0xd4, 0xe1, 0x9d, 0x16, 0xde, 0x2e, 0x40,
+	0x18, 0x50, 0x60, 0x33, 0x86, 0x3b, 0x74, 0x1c, 0x94, 0x26, 0x65, 0x85, 0x8c, 0x04, 0xa5, 0x77,
+	0x77, 0x4a, 0xe2, 0xd1, 0x10, 0x7f, 0xd0, 0x02, 0x80, 0xe9, 0x10, 0xcc, 0x49, 0x4b, 0xe4, 0x30,
+	0x2b, 0x73, 0x98, 0xdb, 0xdd, 0x29, 0x45, 0xac, 0xc6, 0x84, 0xbf, 0xaf, 0x73, 0x6d, 0x13, 0xa6,
+	0x0c, 0xd2, 0xa6, 0x2e, 0x27, 0x8e, 0x41, 0xee, 0xf4, 0x89, 0xcb, 0x91, 0x1a, 0xa9, 0xb7, 0xec,
+	0x86, 0x48, 0x85, 0x8f, 0x06, 0x15, 0x96, 0x8d, 0x10, 0x94, 0x55, 0x85, 0x71, 0x1b, 0xbb, 0xee,
+	0x5d, 0xe6, 0xb4, 0xbc, 0xd2, 0x1b, 0xc3, 0x67, 0xed, 0x32, 0x4c, 0x86, 0x04, 0x76, 0x77, 0x1b,
+	0xe5, 0x86, 0x6d, 0xa6, 0xc8, 0xf6, 0x52, 0xe3, 0xed, 0x15, 0xd2, 0x69, 0xef, 0x42, 0xb6, 0x21,
+	0x0a, 0xfe, 0x7a, 0x42, 0xab, 0x03, 0xf8, 0xde, 0x45, 0x5c, 0x79, 0x50, 0xfa, 0x41, 0xff, 0x1b,
+	0x62, 0x8b, 0x4e, 0x41, 0x16, 0x9b, 0x26, 0x71, 0xdd, 0x4d, 0xce, 0x6e, 0x13, 0xcb, 0x77, 0x9c,
+	0xf1, 0x6c, 0x1b, 0xc2, 0xa4, 0x2d, 0xc3, 0x64, 0x83, 0xb5, 0x59, 0x9f, 0x07, 0x11, 0xc6, 0x6f,
+	0xf7, 0x0a, 0x3e, 0x26, 0x21, 0x13, 0xf8, 0xb0, 0xbb, 0xdb, 0x1a, 0x86, 0xe9, 0x15, 0x59, 0x1e,
+	0x31, 0x94, 0xaf, 0xe7, 0xe2, 0xeb, 0x30, 0x15, 0xa5, 0x38, 0x64, 0x55, 0x42, 0x42, 0x25, 0x42,
+	0xa8, 0xfd, 0x99, 0x84, 0xe9, 0x9b, 0x76, 0x2b, 0x16, 0x78, 0xe8, 0x77, 0xe2, 0xdf, 0xf9, 0x15,
+	0xd6, 0x88, 0x66, 0x04, 0x42, 0x71, 0x3c, 0x2e, 0x14, 0xa1, 0x34, 0x5c, 0x3c, 0x94, 0x34, 0x0c,
+	0xe5, 0x40, 0x8d, 0xcb, 0x41, 0x44, 0x00, 0x66, 0xf7, 0x0a, 0xc0, 0x70, 0xe8, 0x2f, 0x1e, 0x6a,
+	0xe8, 0x87, 0x83, 0xae, 0xc6, 0x07, 0x3d, 0x32, 0xda, 0xf9, 0xc8, 0x68, 0x7b, 0x13, 0x7d, 0x12,
+	0xa0, 0x2f, 0xf3, 0x2a, 0x27, 0x7a, 0x52, 0x5e, 0x76, 0xc2, 0xb7, 0xd4, 0xb9, 0x36, 0x0d, 0x53,
+	0xd1, 0xb4, 0x8b, 0x16, 0xfa, 0x3c, 0x01, 0x33, 0x9e, 0xed, 0x6d, 0xbf, 0xe4, 0x07, 0x94, 0x63,
+	0xd8, 0x25, 0xc9, 0xbd, 0x5d, 0x22, 0x5a, 0xd7, 0x22, 0x77, 0x37, 0x63, 0x5d, 0x94, 0xb1, 0xc8,
+	0xdd, 0xc0, 0x2b, 0xfa, 0x1f, 0xe4, 0x4d, 0x66, 0x6d, 0x51, 0xa7, 0x17, 0xc2, 0xbc, 0x32, 0x4d,
+	0xf9, 0xf6, 0x00, 0xaa, 0xcd, 0xc0, 0x91, 0x78, 0x48, 0x22, 0xd4, 0x2a, 0xe4, 0xae, 0x12, 0x1e,
+	0xed, 0x98, 0xb9, 0xc8, 0xbf, 0xa1, 0xec, 0x1f, 0xe2, 0x1f, 0x8d, 0x43, 0x97, 0x34, 0xda, 0xd2,
+	0x44, 0xc0, 0xda, 0x65, 0xc8, 0x0e, 0xf1, 0xa2, 0x6f, 0xcf, 0xc1, 0xa8, 0x48, 0xb4, 0xc4, 0x67,
+	0x6a, 0xc7, 0xf6, 0xc9, 0xbf, 0xc4, 0x4a, 0x90, 0x56, 0x81, 0xa3, 0xcb, 0x98, 0x9b, 0x1d, 0xdf,
+	0x83, 0x1b, 0x50, 0xe6, 0x41, 0xa1, 0x2d, 0xb7, 0x90, 0x28, 0x2b, 0x15, 0xc5, 0x10, 0x5b, 0x6d,
+	0x05, 0x50, 0x0c, 0x29, 0xc8, 0x16, 0x60, 0x4c, 0xf8, 0xf1, 0x90, 0x07, 0xb0, 0x79, 0xa8, 0xf9,
+	0x05, 0x80, 0xb0, 0xcf, 0x10, 0x40, 0x6a, 0xed, 0x86, 0x71, 0xbd, 0xde, 0xc8, 0x8f, 0x88, 0xfd,
+	0x95, 0xd5, 0xc6, 0xea, 0xc6, 0x6a, 0x3e, 0x81, 0xd2, 0xa0, 0x2c, 0x63, 0x2b, 0x9f, 0x9c, 0xd7,
+	0x01, 0xc2, 0x5e, 0x41, 0x19, 0x48, 0xdf, 0x5c, 0x7b, 0x6b, 0xed, 0xc6, 0xad, 0xb5, 0xfc, 0x08,
+	0x1a, 0x87, 0xd1, 0xeb, 0xf5, 0x86, 0x40, 0x03, 0xa4, 0xde, 0x58, 0x95, 0xfb, 0x64, 0xed, 0xfe,
+	0x18, 0x64, 0x04, 0xdf, 0x3a, 0x71, 0x06, 0xd4, 0x24, 0xc8, 0x82, 0xf1, 0x40, 0x6a, 0x91, 0xb6,
+	0x4f, 0x6c, 0x31, 0xa1, 0x57, 0xcb, 0x07, 0x62, 0x44, 0x75, 0xe6, 0x3e, 0xf8, 0xe9, 0xe9, 0x17,
+	0xc9, 0x59, 0x6d, 0x5a, 0x7c, 0x9b, 0xe0, 0x3e, 0xef, 0xe8, 0x8e, 0xff, 0x7e, 0x29, 0x31, 0x8f,
+	0x9a, 0x30, 0x26, 0xf5, 0x13, 0x95, 0xf6, 0x71, 0x14, 0xd5, 0x6d, 0xf5, 0xe4, 0xcb, 0x01, 0x82,
+	0xe6, 0xb8, 0xa4, 0x39, 0xa2, 0xe5, 0x86, 0x34, 0x72, 0x96, 0x05, 0xc7, 0x16, 0xa4, 0x3c, 0x71,
+	0x44, 0xe5, 0xfd, 0x7d, 0x84, 0xda, 0xab, 0x16, 0x0f, 0x40, 0x08, 0x1a, 0x55, 0xd2, 0x1c, 0xd5,
+	0xa6, 0xa2, 0x34, 0xac, 0xcf, 0x05, 0x0f, 0x81, 0xb4, 0x5f, 0x6b, 0x74, 0x6a, 0xdf, 0x21, 0x8e,
+	0xf6, 0xa8, 0x5a, 0x3a, 0x08, 0x22, 0xa8, 0x66, 0x25, 0x55, 0x1e, 0xe5, 0x82, 0x8f, 0x3a, 0x57,
+	0x7f, 0x8f, 0xb6, 0xde, 0x47, 0x77, 0x00, 0xc2, 0x61, 0x45, 0xff, 0xdd, 0xaf, 0x81, 0xe2, 0x12,
+	0xaa, 0x6a, 0xff, 0x80, 0x8a, 0x64, 0x50, 0x8d, 0xf1, 0x89, 0x9b, 0x7d, 0x94, 0x80, 0xdc, 0xde,
+	0xc9, 0x43, 0x95, 0x97, 0x7a, 0x8c, 0xe9, 0x85, 0x7a, 0xe6, 0x15, 0x90, 0x82, 0x5f, 0x93, 0xfc,
+	0x73, 0xb5, 0x63, 0x21, 0x7f, 0x20, 0x08, 0x41, 0x20, 0xcb, 0x3f, 0x28, 0x0f, 0xaf, 0xfe, 0x9e,
+	0x44, 0xdf, 0x25, 0x20, 0x4d, 0x2d, 0x77, 0x01, 0xdb, 0x14, 0x9d, 0xb8, 0x42, 0x5c, 0xd3, 0xa1,
+	0xb6, 0xf8, 0x1e, 0x2d, 0x6f, 0x39, 0xac, 0x57, 0x0e, 0xbf, 0x4f, 0xc5, 0xa8, 0xaf, 0x30, 0x8b,
+	0x63, 0x93, 0x97, 0xd7, 0x84, 0x08, 0x9f, 0xeb, 0x70, 0x6e, 0xbb, 0x4b, 0xba, 0xde, 0xa6, 0xbc,
+	0xd3, 0x6f, 0x56, 0x4d, 0xd6, 0xd3, 0xdb, 0x6c, 0x41, 0x06, 0xe7, 0x7a, 0x23, 0xa0, 0x9b, 0x1d,
+	0xcc, 0xdb, 0x0e, 0xee, 0xcd, 0x37, 0x20, 0x73, 0xfd, 0xda, 0x46, 0xb9, 0x41, 0x4d, 0x62, 0xb9,
+	0x04, 0xfd, 0xff, 0x10, 0x67, 0xf5, 0x66, 0x97, 0x35, 0xf5, 0x1e, 0xa6, 0x96, 0xde, 0xb8, 0xb6,
+	0xb2, 0xba, 0xb6, 0xbe, 0x5a, 0x4b, 0x0d, 0x16, 0xab, 0xe7, 0xab, 0xe7, 0xd5, 0xd3, 0x30, 0x23,
+	0x1c, 0x2d, 0xe9, 0x7a, 0x97, 0x99, 0xb8, 0xdb, 0x61, 0x2e, 0x5f, 0xba, 0x74, 0xfe, 0xd2, 0x22,
+	0xca, 0xfe, 0x7a, 0xef, 0xfe, 0xb3, 0x7b, 0x5f, 0x3f, 0xff, 0xf4, 0xc1, 0xb3, 0xaf, 0x3e, 0x3c,
+	0x08, 0xf6, 0xcd, 0x27, 0xbf, 0x3d, 0xf8, 0xd2, 0x87, 0x9d, 0x81, 0x59, 0x1f, 0xe6, 0xe7, 0xa3,
+	0x6a, 0x32, 0xd6, 0xad, 0x52, 0x86, 0xb2, 0xcf, 0x1f, 0x7e, 0xff, 0xf4, 0xdb, 0x8f, 0x3d, 0xdc,
+	0xfc, 0xd9, 0xa5, 0xd3, 0xf0, 0x1f, 0x80, 0x26, 0xc1, 0x0e, 0x71, 0xea, 0x7d, 0xde, 0x41, 0x33,
+	0x70, 0x04, 0x46, 0xc5, 0xd9, 0xf9, 0x94, 0x67, 0xad, 0x29, 0x6f, 0xde, 0xda, 0xa8, 0xe5, 0x21,
+	0xb7, 0x07, 0x36, 0xf2, 0xe3, 0xe3, 0x62, 0xe2, 0xd1, 0xe3, 0x62, 0xe2, 0x97, 0xc7, 0xc5, 0xc4,
+	0x67, 0x4f, 0x8a, 0x23, 0x8f, 0x9e, 0x14, 0x47, 0x7e, 0x7e, 0x52, 0x1c, 0x79, 0xe7, 0xd2, 0xab,
+	0xe4, 0xe1, 0x85, 0x9f, 0x27, 0x97, 0x07, 0x8b, 0xcd, 0x94, 0xfc, 0xd5, 0x70, 0xe1, 0xef, 0x00,
+	0x00, 0x00, 0xff, 0xff, 0x18, 0xdc, 0xa2, 0xdf, 0xbd, 0x0c, 0x00, 0x00,
 }
 
 func (m *User) Marshal() (dAtA []byte, err error) {
@@ -1446,10 +1447,10 @@ func (m *LoginReply) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Token) > 0 {
-		i -= len(m.Token)
-		copy(dAtA[i:], m.Token)
-		i = encodeVarintUser(dAtA, i, uint64(len(m.Token)))
+	if len(m.AccessToken) > 0 {
+		i -= len(m.AccessToken)
+		copy(dAtA[i:], m.AccessToken)
+		i = encodeVarintUser(dAtA, i, uint64(len(m.AccessToken)))
 		i--
 		dAtA[i] = 0x12
 	}
@@ -1483,10 +1484,10 @@ func (m *LogoutRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 	_ = i
 	var l int
 	_ = l
-	if len(m.Token) > 0 {
-		i -= len(m.Token)
-		copy(dAtA[i:], m.Token)
-		i = encodeVarintUser(dAtA, i, uint64(len(m.Token)))
+	if len(m.AccessToken) > 0 {
+		i -= len(m.AccessToken)
+		copy(dAtA[i:], m.AccessToken)
+		i = encodeVarintUser(dAtA, i, uint64(len(m.AccessToken)))
 		i--
 		dAtA[i] = 0x12
 	}
@@ -1763,10 +1764,10 @@ func (m *UpdatePasswordRequest) MarshalToSizedBuffer(dAtA []byte) (int, error) {
 		i--
 		dAtA[i] = 0x1a
 	}
-	if len(m.OldPassword) > 0 {
-		i -= len(m.OldPassword)
-		copy(dAtA[i:], m.OldPassword)
-		i = encodeVarintUser(dAtA, i, uint64(len(m.OldPassword)))
+	if len(m.Password) > 0 {
+		i -= len(m.Password)
+		copy(dAtA[i:], m.Password)
+		i = encodeVarintUser(dAtA, i, uint64(len(m.Password)))
 		i--
 		dAtA[i] = 0x12
 	}
@@ -2079,7 +2080,7 @@ func (m *LoginReply) Size() (n int) {
 	if l > 0 {
 		n += 1 + l + sovUser(uint64(l))
 	}
-	l = len(m.Token)
+	l = len(m.AccessToken)
 	if l > 0 {
 		n += 1 + l + sovUser(uint64(l))
 	}
@@ -2095,7 +2096,7 @@ func (m *LogoutRequest) Size() (n int) {
 	if m.Id != 0 {
 		n += 1 + sovUser(uint64(m.Id))
 	}
-	l = len(m.Token)
+	l = len(m.AccessToken)
 	if l > 0 {
 		n += 1 + l + sovUser(uint64(l))
 	}
@@ -2224,7 +2225,7 @@ func (m *UpdatePasswordRequest) Size() (n int) {
 	if l > 0 {
 		n += 1 + l + sovUser(uint64(l))
 	}
-	l = len(m.OldPassword)
+	l = len(m.Password)
 	if l > 0 {
 		n += 1 + l + sovUser(uint64(l))
 	}
@@ -3149,7 +3150,7 @@ func (m *LoginReply) Unmarshal(dAtA []byte) error {
 			iNdEx = postIndex
 		case 2:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Token", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field AccessToken", wireType)
 			}
 			var stringLen uint64
 			for shift := uint(0); ; shift += 7 {
@@ -3177,7 +3178,7 @@ func (m *LoginReply) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Token = string(dAtA[iNdEx:postIndex])
+			m.AccessToken = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
@@ -3250,7 +3251,7 @@ func (m *LogoutRequest) Unmarshal(dAtA []byte) error {
 			}
 		case 2:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Token", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field AccessToken", wireType)
 			}
 			var stringLen uint64
 			for shift := uint(0); ; shift += 7 {
@@ -3278,7 +3279,7 @@ func (m *LogoutRequest) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Token = string(dAtA[iNdEx:postIndex])
+			m.AccessToken = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex
@@ -4125,7 +4126,7 @@ func (m *UpdatePasswordRequest) Unmarshal(dAtA []byte) error {
 			iNdEx = postIndex
 		case 2:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field OldPassword", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field Password", wireType)
 			}
 			var stringLen uint64
 			for shift := uint(0); ; shift += 7 {
@@ -4153,7 +4154,7 @@ func (m *UpdatePasswordRequest) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.OldPassword = string(dAtA[iNdEx:postIndex])
+			m.Password = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		case 3:
 			if wireType != 2 {
