@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
 // - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.12.3
+// - protoc             v3.18.1
 // source: api/micro/user/v1/user.proto
 
 package v1
@@ -23,12 +23,17 @@ const _ = grpc.SupportPackageIsVersion7
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
 type UserServiceClient interface {
 	// auth
+	// 注册
 	Register(ctx context.Context, in *RegisterRequest, opts ...grpc.CallOption) (*RegisterReply, error)
+	// 登录
 	Login(ctx context.Context, in *LoginRequest, opts ...grpc.CallOption) (*LoginReply, error)
+	// 退出
 	Logout(ctx context.Context, in *LogoutRequest, opts ...grpc.CallOption) (*LogoutReply, error)
-	// user
+	// 获取用户信息
 	GetUser(ctx context.Context, in *GetUserRequest, opts ...grpc.CallOption) (*GetUserReply, error)
+	// 更新用户信息
 	UpdateUser(ctx context.Context, in *UpdateUserRequest, opts ...grpc.CallOption) (*UpdateUserReply, error)
+	// 修改用户密码
 	UpdatePassword(ctx context.Context, in *UpdatePasswordRequest, opts ...grpc.CallOption) (*UpdatePasswordReply, error)
 }
 
@@ -99,12 +104,17 @@ func (c *userServiceClient) UpdatePassword(ctx context.Context, in *UpdatePasswo
 // for forward compatibility
 type UserServiceServer interface {
 	// auth
+	// 注册
 	Register(context.Context, *RegisterRequest) (*RegisterReply, error)
+	// 登录
 	Login(context.Context, *LoginRequest) (*LoginReply, error)
+	// 退出
 	Logout(context.Context, *LogoutRequest) (*LogoutReply, error)
-	// user
+	// 获取用户信息
 	GetUser(context.Context, *GetUserRequest) (*GetUserReply, error)
+	// 更新用户信息
 	UpdateUser(context.Context, *UpdateUserRequest) (*UpdateUserReply, error)
+	// 修改用户密码
 	UpdatePassword(context.Context, *UpdatePasswordRequest) (*UpdatePasswordReply, error)
 	mustEmbedUnimplementedUserServiceServer()
 }
