@@ -10,11 +10,11 @@ import (
 	"github.com/swaggo/gin-swagger/swaggerFiles"
 
 	// import swagger handler
-	// _ "github.com/go-microservice/ins-api/docs" // docs is generated by Swag CLI, you have to import it.
+	// _ "github.com/go-microservice/chatgram/docs" // docs is generated by Swag CLI, you have to import it.
 
 	"github.com/go-eagle/eagle/pkg/app"
 	"github.com/go-eagle/eagle/pkg/middleware"
-	"github.com/go-microservice/ins-api/internal/handler"
+	"github.com/go-microservice/chatgram/internal/handler"
 )
 
 // Load loads the middlewares, routes, handlers.
